+++ luogu/P1478.cpp
@@ -24,26 +24,26 @@
 
 using namespace std;
 
-int main()
+long main()
 {
     // 苹果数量，力气
-    int n, s;
+    long n, s;
     // 椅子高度，陶陶最大伸长高度
-    int a, b;
+    long a, b;
     // 苹果列表，分别存储其高度与所费力气
     vector<pair<int, int>> apples;
 
     cin >> n >> s >> a >> b;
-    for (int i = 0; i < n; ++i)
+    for (long i = 0; i < n; ++i)
     {
         // 苹果高度，花费力气
-        int x, y;
+        long x, y;
         cin >> x >> y;
         apples.push_back(make_pair(x, y));
     }
 
     // 化简数值，将最大高度加起来
-    int height = a + b;
+    long height = a + b;
 
     // 根据花费的力气进行升序排列
     sort(apples.begin(), apples.end(), [](const pair<int, int> &l, const pair<int, int> &r) -> bool
@@ -57,7 +57,7 @@ int main()
          });
 
     // 过一遍苹果列表，统计能摘的苹果数
-    int count = 0;
+    long count = 0;
     for (auto &i : apples)
     {
         if (height >= i.first && s >= i.second)
